style: renaming of folder
