+++ src/app/tree/spec/groups/bimodal.ts
