+++ src/app/tree/spec/legends.ts
